Creating Coupon with refrrence to Company id works now !

+++ src/main/java/com/Spring/CouponSystem/Beans/Repository/CompanyRepo.java
@@ -15,14 +15,13 @@ import com.Spring.CouponSystem.Beans.Customer;
 import com.Spring.CouponSystem.Beans.Enum.CouponType;
 
 @Repository
-public interface CompanyRepo extends JpaRepository<Company, Long> {
+public interface CompanyRepo extends JpaRepository<Company, Integer> {
 
 	@Query("Select c from Company c where c.id = :id")
 	Company findById(int id);
 
 //	Company findCompanyByCompNameAndPassword(String user, String password);
-	
-	
+
 //	@Query("select c from Coupon c where c.company = ?1")
 //	List<Coupon> findCompanyCoupons(Company company);
 //	
@@ -60,5 +59,5 @@ public interface CompanyRepo extends JpaRepository<Company, Long> {
 
 	@Query("select c from Coupon c where c.type = :type")
 	List<Coupon> findCouponsByType(CouponType type);
-	
+
 }
